% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycorr1.R
\name{panel.cor}
\alias{panel.cor}
\title{Make correlation plot for plot of class "mycor"}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{digits}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used.}

\item{prefix}{a character vector}

\item{cex.cor}{a numeric variable}
}
\description{
Make correlation plot for plot of class "mycor"
}
