% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel.hist}
\alias{panel.hist}
\title{Make plot with histogram for plot of class "mycor"}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Make plot with histogram for plot of class "mycor"
}

