% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycorr1.R
\name{summary.mycor}
\alias{summary.mycor}
\title{Summarizing function for class "mycor"}
\usage{
\method{summary}{mycor}(object, ...)
}
\arguments{
\item{object}{an object of class "mycor", a result of a call to \code{\link{mycor}}.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Summarizing function for class "mycor"
}
\examples{
out=mycor(iris)
summary(out)
}
