% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycorr1.R
\name{panel.hist}
\alias{panel.hist}
\title{Make plot with histogram for plot of class "mycor"}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Make plot with histogram for plot of class "mycor"
}
