% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.mycor}
\alias{print.mycor}
\title{Print function for class "mycor"}
\usage{
\method{print}{mycor}(x, ...)
}
\arguments{
\item{x}{an object of class "mycor", a result of a call to \code{\link{mycor}}.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Print function for class "mycor"
}
\examples{
out=mycor(iris)
print(out)
}

